fix-review

+++ src/main/java/com/anhoang/shopping/online/service/ReviewService.java
@@ -4,15 +4,11 @@ import com.anhoang.shopping.online.model.Review;
 import com.anhoang.shopping.online.model.Product;
 import com.anhoang.shopping.online.model.UserDtls;
 
-import java.time.LocalDateTime;
 import java.util.List;
 
 public interface ReviewService {
     Review addReview(Product product, UserDtls user, int rating, String comment);
     List<Review> getReviewsByProduct(Product product);
-    void deleteReview(Integer reviewId);
-
-    Review getReviewById(Integer reviewId);
-
-
+    void deleteReview(Long reviewId);               // <-- Long
+    Review getReviewById(Long reviewId);           // <-- Long
 }
